requirements.txt
blinker==1.9.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
Flask==3.1.3
idna==3.15
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
PyYAML==6.0.3
requests==2.34.1
tzdata==2026.2
urllib3==2.7.0
Werkzeug==3.1.8
